+++ mpltex/styles.py
@@ -9,11 +9,13 @@ Toolkits and constants for plotting styles.
 
 import itertools
 
-from .colors import almost_black, brewer_set1
+from .colors import almost_black, tableau_10
 
-__all__ = ['lines', 'markers', 'nextlinestyle', 'linestyle_generator']
+__all__ = ['colors', 'lines',
+           'markers', 'markersh', 'marker_types',
+           'linestyle_generator', 'linestyles']
 
-_colors = brewer_set1
+_colors = tableau_10
 colors = itertools.cycle(_colors)
 
 _lines = ['-', '--', '-.', ':']
@@ -28,6 +30,15 @@ markersh = itertools.cycle(_markersh)
 
 _marker_types = [False, True]  # True for hollow markers
 marker_types = itertools.cycle(_marker_types)
+
+
+def linestyles(colors=_colors, lines=_lines,
+               markers=_markers, hollow_styles=_marker_types):
+    """
+    Short name for ``linestyle_generator``.
+    To be compatible with previous versions.
+    """
+    return linestyle_generator(colors, lines, markers, hollow_styles)
 
 
 def linestyle_generator(colors=_colors, lines=_lines,
@@ -106,60 +117,3 @@ def linestyle_generator(colors=_colors, lines=_lines,
                    'marker': marker, 'mew': mew, 'mec': mec, 'mfc': mfc}
         else:
             yield {}
-
-
-def nextlinestyle(no_line=False, is_line=True, is_marker=True, is_hollow=True):
-    """
-    Generate a dict for configuring plot line style.
-
-    NOTE: this function has a serious problem.
-    As long as mpltex is imported,
-    the beginning of the sequence cannot be configured later.
-    This function is obsolete since version 0.2, please use linestyle_generator instead.
-
-    The default line style is markers linked by lines, both styles are cycled,
-    and hollow markers are included.
-    Disable cycling one style by set the corresponding flag to False.
-
-    :param no_line: if True, only markers, no line segments
-    :type no_line: bool
-    :param is_line: if True, cycle line styles. No effect when no_line=True.
-    :type is_line: bool
-    :param is_marker: if True, cycle markers.
-    :type is_marker: bool
-    :param is_hollow: if True, cycle markers including hollow styles.
-    :param is_hollow: bool
-    :return: dict of parameters of linestyle
-    :rtype: dict
-    """
-    color = colors.next()
-    linestyle = lines.next()
-    if(is_hollow):
-        marker = markersh.next()
-    else:
-        marker = markers.next()
-    marker_type = marker_types.next()
-    if marker_type:  # hollow mark
-        mew = 1
-        mec = color
-        mfc = 'None'
-    else:
-        mew = 1
-        mec = color
-        mfc = color
-
-    linestyles = {}
-    if (is_line):
-        linestyles['linestyle'] = linestyle
-    if (no_line):  # overwrite is_line if no_line is True
-        linestyles['linestyle'] = ''
-    if (is_marker):
-        linestyles['marker'] = marker
-        linestyles['mew'] = 0  # make the marker no edge
-    if (is_hollow):
-        linestyles['color'] = color
-        linestyles['marker'] = marker
-        linestyles['mew'] = mew
-        linestyles['mec'] = mec
-        linestyles['mfc'] = mfc
-    return linestyles